pyproject.toml
[project]
name = "can-i-use-terminal-github-io"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "hkalbasi", email = "hamidrezakalbasi@protonmail.com" }
]
dependencies = [
    "python-liquid>=1.12.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
